refactor(post): 도메인을 post 패키지로 위치 변경
