+++ src/main/java/com/coredisc/domain/member/Member.java
@@ -1,8 +1,8 @@
 package com.coredisc.domain.member;
 
 import com.coredisc.domain.Comment;
-import com.coredisc.domain.Post;
-import com.coredisc.domain.PostLike;
+import com.coredisc.domain.post.Post;
+import com.coredisc.domain.post.PostLike;
 import com.coredisc.domain.common.BaseEntity;
 import com.coredisc.domain.common.enums.OauthType;
 import com.fasterxml.jackson.annotation.JsonProperty;
