+++ src/main/java/com/coredisc/domain/post/PostAnswer.java
@@ -1,7 +1,8 @@
-package com.coredisc.domain;
+package com.coredisc.domain.post;
 
 
 
+import com.coredisc.domain.TodayQuestion;
 import com.coredisc.domain.common.enums.AnswerType;
 import com.coredisc.domain.common.BaseEntity;
 import jakarta.persistence.*;
