+++ src/main/java/com/coredisc/domain/post/PostAnswerImage.java
@@ -1,4 +1,4 @@
-package com.coredisc.domain;
+package com.coredisc.domain.post;
 
 
 import com.coredisc.domain.common.BaseEntity;
